+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `@lumx/react`:
     -    `Popover`: added `closeMode` to `"unmount"` (default) or `"hide"` the popover when closed
+-   `@lumx/core`:
+    -   Add action area CSS utils
 
 ## [4.6.0][] - 2026-03-03